Keep initial drag position when dragging
When dragging an object, the dragged element shouldn't be centered
around the touch position, but should be dragged relative to where we
touched.

For this we need to know where we started touching, and also where the
original element was in relation to the screen.

This only applies to default clones, and not for custom elements, which
must be placed manually.